api/requirements.txt
altgraph==0.17.2
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
autopep8==1.7.0
blabel==0.1.6
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect2==0.7.0
cx-Freeze==6.11.1
cx-Oracle==8.3.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
drf-yasg2==1.19.4
fonttools==4.37.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lief==0.12.1
lxml==4.9.1
MarkupSafe==2.1.1
oracledb==1.1.0
oscrypto==1.3.0
packaging==21.3
pefile==2022.5.30
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.3.0
Pygments==2.13.0
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PyJWT==2.5.0
pynput==1.7.6
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.13.0
pyStrich==0.8
python-bidi==0.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.1.1
toml==0.10.2
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.12
weasyprint==56.1
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
zopfli==0.2.1
